allow to give a table of elements that should/should not be listed
